Only allow punct delimiter for subpattern
The experimental feature that allows wildcard subpatterns in finding
Unicode properties, is supposed to only allow ASCII punctuation for
delimitters.  But if you preceded the delimitter by a backslash, the
check was skipped.  This fixes that.

It may be that we will eventually want to loosen the restriction and
allow a wider range of delimiters.  But until we have valid use-cases
that would encourage us to do so, I don't want to get into supporting
stuff that we might later regret, such as invisible characters for
delimitters.  This feature is just a nicety, and not really required for
programs to work, so I don't view it as necessary to be as general as
possible.